task3/requirements.txt
SQLAlchemy==2.0.30
alembic==1.13.1
fastapi==0.110.0
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pydantic==2.6.4
pytest==8.1.1
redis==5.0.4
requests==2.31.0
httpx==0.27.0
uvicorn==0.29.0
